% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{arg_check_data}
\alias{arg_check_data}
\title{(Internal) Check data frames provided as inputs for basic requirements.}
\usage{
arg_check_data(arg_data, arg_name)
}
\arguments{
\item{arg_data}{A data frame passed as an argument to be checked.}

\item{arg_name}{A character of length 1 indicating the name of the argument.}
}
\value{
NULL. Throws an error if the data frame doesn't pass the checks.
}
\description{
(Internal) Check data frames provided as inputs for basic requirements.
}
