% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_join}
\alias{check_join}
\title{(Internal) Argument checking for joining two data frames}
\usage{
check_join(
  x,
  y,
  join,
  arg_x = caller_arg(x),
  arg_y = caller_arg(y),
  arg_j = caller_arg(join),
  call = caller_env()
)
}
\arguments{
\item{x}{A data frame.}

\item{y}{A data frame.}

\item{join}{A \code{dplyr::join_by} join specification or NULL.}

\item{arg_x}{The calling function's argument corresponding to \code{x}.}

\item{arg_y}{The calling function's argument corresponding to \code{y}.}

\item{arg_j}{The calling function's argument corresponding to \code{join}.}

\item{call}{The calling function's environment.}
}
\value{
Error if argument conditions are not met, NULL otherwise.
}
\description{
(Internal) Argument checking for joining two data frames
}
