% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_name}
\alias{check_name}
\title{(Internal) Argument checking for names.}
\usage{
check_name(x, arg = caller_arg(x), call = caller_env())
}
\arguments{
\item{x}{A character of length 1.}

\item{arg}{The calling function's argument corresponding to \code{x}.}

\item{call}{The calling function's environment.}
}
\value{
Error if argument conditions are not met, NULL otherwise.
}
\description{
(Internal) Argument checking for names.
}
