% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformalScore.R
\name{conformalScore}
\alias{conformalScore}
\title{conformalScore}
\usage{
conformalScore(fit, fusion_choice = "early", conf_level = 0.95)
}
\arguments{
\item{fit}{A fitted model object. For early and late fusion, this should be an
\emph{IntegratedLearner} object. For intermediate fusion, this should be a
\emph{BayesCOOP} object.}

\item{fusion_choice}{Fusion strategy to use. One of \code{"early"}, \code{"intermediate"}, or
\code{"late"}. Default is \code{"early"}.}

\item{conf_level}{Desired marginal confidence level. Default is \code{0.95}.}
}
\value{
A list with components:
\item{d}{Conformal cutoff corresponding to \code{conf_level}.}
\item{conf_level}{Confidence level used.}
\item{fusion_choice}{Fusion strategy used.}
\item{fit}{Fitted model object.}
\item{conformal_scores}{Vector of conformal scores.}
}
\description{
Computes conformal scores and the associated cutoff for a fitted multiview model
under early, late, or intermediate fusion.
}
