% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coracle-package.R
\docType{package}
\name{coracle-package}
\alias{coracle}
\alias{coracle-package}
\title{coracle: Correlations for Columns of Tidy Data}
\description{
Provices functions for correlating data stored in columns.
}
\author{
\strong{Maintainer}: Ray Krajci \email{rdk54@case.edu}

}
\keyword{internal}
