% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coracle.R
\name{Coracle}
\alias{Coracle}
\title{Coracle}
\usage{
Coracle(
  fit = NULL,
  fit_coop = NULL,
  data_calib,
  data_valid,
  fusion_choice = c("late", "early", "intermediate"),
  conf_level = 0.95
)
}
\arguments{
\item{fit}{A fitted \emph{IntegratedLearner} object from early/late fusion. See \emph{IntegratedLearner} for details. Default: NULL.}

\item{fit_coop}{A fitted \emph{BayesCOOP} object for intermediate fusion. See \emph{BayesCOOP} for details. Default: NULL.}

\item{data_calib}{Calibration data list containing feature_table, feature_metadata,
and sample_metadata. See \emph{IntegratedLearner} for details.}

\item{data_valid}{Validation data list containing feature_table, feature_metadata,
and sample_metadata. See \emph{IntegratedLearner} for details.}

\item{fusion_choice}{Fusion strategy to use. One of \code{"early"}, \code{"intermediate"}, or
\code{"late"}. Default is \code{"early"}.}

\item{conf_level}{Desired marginal confidence level. Default: 0.95.}
}
\value{
A list with components:
\item{df}{A dataframe containing the Coracle output for different fusion choices.}
\item{coverage}{A list containing the coverage for different fusion choices.}
\item{conf_level}{Desired marginal confidence level.}
}
\description{
Conformalized multiview prediction under early, late, or intermediate fusion.
Produces finite-sample valid prediction intervals and empirical coverage
on a validation set.
}
