% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coracle_data.R
\name{coracle_data}
\alias{coracle_data}
\title{R6 Class Representing Data for Correlation with the \code{coracle} package}
\value{
A \code{coracle_data} root node.

Data as a \code{data.frame}.

List of \code{coracle_data} leaf nodes.

List of \code{coracle_data} leaf nodes with data valid for correlation.

List of \code{coracle_data} leaf nodes with data invalid for correlation.
}
\description{
A \code{coracle_data} nodes stores data for correlation in a tree.
}
\details{
Branches are based on grouping variables. Data is stored as chunks in the leaf nodes of the tree.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{children}}{List of child \code{coracle_data} node(s).}

\item{\code{chunk}}{\code{data.frame} of data for correlation.}

\item{\code{grps_vals}}{Named list of grouping column(s) and chunk value(s).}

\item{\code{chunk_flags}}{Boolean indication of whether this node is valid for correlation.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-coracle_data-new}{\code{coracle_data$new()}}
\item \href{#method-coracle_data-clone}{\code{coracle_data$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-coracle_data-new"></a>}}
\if{latex}{\out{\hypertarget{method-coracle_data-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{coracle_data} node
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{coracle_data$new(
  data = NULL,
  grps = NULL,
  join = NULL,
  vals = NULL,
  ...,
  labl_cols = "cols",
  labl_vals = "vals",
  parent = NULL,
  call = caller_env()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A \code{data.frame} of data to prepare for correlation.}

\item{\code{grps}}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> expressions identifying grouping column(s)}

\item{\code{join}}{A <\code{\link[=dplyr_tidy_select]{tidy-select}}> expression identifying one joining column.}

\item{\code{vals}}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> expressions identifying value column(s).}

\item{\code{...}}{Unused}

\item{\code{labl_cols}}{(Optional) Character scalar column name for long-format groups.}

\item{\code{labl_vals}}{(Optional) Character scalar column name for long-format values.}

\item{\code{parent}}{(Optional) A \code{coracle_data} node.}

\item{\code{call}}{Caller environment for error reporting.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-coracle_data-clone"></a>}}
\if{latex}{\out{\hypertarget{method-coracle_data-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{coracle_data$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
