% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_col.R
\name{corr_col}
\alias{corr_col}
\title{Correlate all columns of two data frames}
\usage{
corr_col(x_data, y_data, ...)
}
\arguments{
\item{x_data}{A data frame with at least one shared index column and at least one column of numeric data.}

\item{y_data}{A data frame with at least one shared index column and at least one column of numeric data.}

\item{...}{}
}
\value{
A data frame of correlation results.
}
\description{
Correlate all columns of two data frames
}
\examples{
corr_col(df1, df2)
}
