% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coracle.R
\name{corr_result}
\alias{corr_result}
\title{(Internal) assemble correlation results in data.frame.}
\usage{
corr_result(
  a,
  b = NULL,
  c = NULL,
  stat_value = NA,
  p = NA,
  n = NA_integer_,
  message = NA_character_,
  method
)
}
\arguments{
\item{a}{A \code{coracle_data} object used in a correlation.}

\item{b}{A \code{coracle_data} object used in a correlation.}

\item{c}{A \code{coracle_data} object used in a correlation.}

\item{stat_value}{The estimated correlation result.}

\item{p}{The p-value of the correlation result.}

\item{n}{The number of observations used in a correlation.}

\item{message}{An informative message about the correlation result.}

\item{method}{The correlation method ("spearman", "pearson", or "kendall")}
}
\value{
A data.frame of correlation results.
}
\description{
(Internal) assemble correlation results in data.frame.
}
