% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_col.R
\name{corr_x}
\alias{corr_x}
\title{(Internal)Mapped correlation function for single data frame (\code{x}) case of \code{corr_col()}}
\usage{
corr_x(x, y, corr_data, method)
}
\arguments{
\item{x}{A column name of \code{x} as a character of length 1}

\item{y}{A column name of \code{x} as a character of length 1}

\item{corr_data}{The data frame prepared from \code{x} for correlation}

\item{method}{(Optional) A character of length 1 indicating which correlation coefficient is to be used: \code{"spearman"} (the default), \code{"pearson"}, or \code{"kendall"}.}
}
\value{
The results of a single correlation as a data frame row.
}
\description{
(Internal)Mapped correlation function for single data frame (\code{x}) case of \code{corr_col()}
}
