% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_col.R
\name{corr_xy}
\alias{corr_xy}
\title{(Internal)Mapped correlation function fort two data frame (\code{x}, \code{y}) case of \code{corr_col()}}
\usage{
corr_xy(var_x, var_y, corr_data, method)
}
\arguments{
\item{var_x}{A column name of \code{x} as a character of length 1}

\item{var_y}{A column name of \code{x} as a character of length 1}

\item{corr_data}{The data frame prepared from \code{x} for correlation}

\item{method}{(Optional) A character of length 1 indicating which correlation coefficient is to be used: \code{"spearman"} (the default), \code{"pearson"}, or \code{"kendall"}.}
}
\value{
The results of a single correlation as a data frame row.
}
\description{
(Internal)Mapped correlation function fort two data frame (\code{x}, \code{y}) case of \code{corr_col()}
}
