% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coracle.R
\name{correlate}
\alias{correlate}
\title{(Internal) Correlate data of \code{coracle_data} objects.}
\usage{
correlate(a, b, method)
}
\arguments{
\item{a}{A \code{coracle_data} object for correlation.}

\item{b}{A \code{coracle_data} object for correlation.}

\item{method}{The correlation method ("spearman", "pearson", or "kendall")}
}
\value{
A \code{corr_result} data.frame.
}
\description{
(Internal) Correlate data of \code{coracle_data} objects.
}
