% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%same_as\%}
\alias{\%same_as\%}
\title{(Internal) Utility function for comparing values, inclusive of NAs}
\usage{
v1 \%same_as\% v2
}
\arguments{
\item{v1}{An atomic value.}

\item{v2}{An atomic value.}
}
\value{
\code{TRUE} where values are the same (including NAs), otherwise \code{FALSE}
}
\description{
(Internal) Utility function for comparing values, inclusive of NAs
}
