% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_col.R
\name{pairwise_corr}
\alias{pairwise_corr}
\title{Internal function for calculation correlations}
\usage{
pairwise_corr(x_var, y_var, data)
}
\arguments{
\item{x_var}{Column name string for correlation}

\item{y_var}{Column name string for correlation}

\item{data}{Data frame with join variables and nested data}
}
\value{
Data frame row of correlation results
}
\description{
Internal function for calculation correlations
}
