% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coracle.R
\name{pairwise_correlation}
\alias{pairwise_correlation}
\title{(Internal) Pairwise correlation of between two \code{coracle_data} objects.}
\usage{
pairwise_correlation(x, y, method)
}
\arguments{
\item{x}{A \code{coracle_data} object for correlation.}

\item{y}{A \code{coracle_data} object for correlation.}

\item{method}{The correlation method ("spearman", "pearson", or "kendall")}
}
\value{
A data.frame of correlation results.
}
\description{
(Internal) Pairwise correlation of between two \code{coracle_data} objects.
}
