% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcor_col.R
\name{pcor_col}
\alias{pcor_col}
\title{Partial correlation of columns for tidy data.}
\usage{
pcor_col(
  x,
  y,
  z,
  ...,
  xy_join = NULL,
  xz_join = NULL,
  x_name = "x",
  y_name = "y",
  z_name = "z",
  method = "spearman"
)
}
\arguments{
\item{x}{A tidy data.frame containing numeric data columns.}

\item{y}{(Optional) A tidy data frame containing numeric data columns for pairwise correlation with \code{x}.}

\item{z}{(Optional) A tidy data frame containing numeric data columns to use as covariates.}

\item{...}{Unused.}

\item{xy_join}{(Optional) A \code{\link[dplyr]{join_by}} join specification for \code{x} and \code{y}.}

\item{xz_join}{(Optional) A \code{\link[dplyr]{join_by}} join specification for \code{x} and \code{z}.}

\item{x_name}{(Optional) A character of length 1 to be used as the first column name.}

\item{y_name}{(Optional) A character of length 1 to be used as the second column name.}

\item{z_name}{(Optional) A character of length 1 to be used as the third column name.}

\item{method}{(Optional) A character of length 1 indicating which correlation coefficient is to be used: \code{"spearman"} (the default), \code{"pearson"}, or \code{"kendall"}.}
}
\value{
A data frame of partial correlation results.
}
\description{
Partial correlation of columns for tidy data.
}
\examples{

# Data frames with a shared column, "i"
x <- data.frame(i = as.character(1:5), ux = 1:5, dx = 5:1)
y <- data.frame(i = as.character(1:5), uy = 1:5, dy = 5:1)
z <- data.frame(i = as.character(1:5), rz = runif(5))
corr_col(x, y, z)

# Data frames without a shared column
x <- data.frame(ix= as.character(1:5), ux = 1:5, dx = 5:1)
y <- data.frame(iy = as.character(1:5), uy = 1:5, dy = 5:1)
z <- data.frame(iz = as.character(1:5), rz = runif(5))
corr_col(x, y,xy_join = dplyr::join_by(ix == iy), xz_join = dplyr::join_by(ix == iz))

# Renaming the outputs
x <- data.frame(i = as.character(1:5), ux = 1:5, dx = 5:1)
y <- data.frame(i = as.character(1:5), uy = 1:5, dy = 5:1)
z <- data.frame(i = as.character(1:5), rz = runif(5))
pcor_col(x,y,z, x_name = "first", y_name = "second", z_name = "third")

}
