% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcor_col.R
\name{pcor_col_x}
\alias{pcor_col_x}
\title{(Internal) Single data frame (\code{x}) case of \code{pcor_col()}}
\usage{
pcor_col_x(x, x_name = "x", ..., method = "spearman")
}
\arguments{
\item{x}{A tidy data frame containing numeric data columns.}

\item{x_name}{(Optional) A character of length 1 to be used as the first column name.}

\item{...}{Unused.}

\item{method}{(Optional) A character of length 1 indicating which correlation coefficient is to be used: \code{"spearman"} (the default), \code{"pearson"}, or \code{"kendall"}.}
}
\value{
A data frame of partial correlation results.
}
\description{
(Internal) Single data frame (\code{x}) case of \code{pcor_col()}
}
