% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{statistic_column}
\alias{statistic_column}
\title{(Internal) Update data frame column with name corresponding to correlation method.}
\usage{
statistic_column(value, method)
}
\arguments{
\item{value}{A numeric or NA value extracted from the correlation results.}

\item{method}{A character of length 1 indicating which correlation coefficient is to be used: \code{"spearman"} (the default), \code{"pearson"}, or \code{"kendall"}.}
}
\value{
A 1x1 data frame of the value in a column named "rho", "cor", or "tau"
}
\description{
(Internal) Update data frame column with name corresponding to correlation method.
}
