% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coracle.R
\name{validate_coracle_input}
\alias{validate_coracle_input}
\title{(Internal) Check non-null inputs for \code{coracle_data} objects}
\usage{
validate_coracle_input(input, arg = caller_arg(input), call = caller_env())
}
\arguments{
\item{input}{The input value.}

\item{arg}{The input argument.}

\item{call}{The caller environment.}
}
\value{
Invisibly.
}
\description{
(Internal) Check non-null inputs for \code{coracle_data} objects
}
