% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_col.R
\name{corr_col}
\alias{corr_col}
\title{Correlate all columns of two data frames}
\usage{
corr_col(x_data, y_data, ..., x_name, y_name, join_vars = NULL)
}
\arguments{
\item{x_data}{A data frame with at least one index column and at least one column of numeric data.}

\item{y_data}{A data frame with at least one index column and at least one column of numeric data.}

\item{...}{Unused}

\item{x_name}{A string identifying the type of data in \code{x_data}}

\item{y_name}{A string identifying the type of data in \code{y_data}}

\item{join_vars}{A named list of column names for joining, passed to \code{dplyr::inner_join()}'s \code{by} argument. See \code{dplyr} reference for information.}
}
\value{
A data frame of correlation results.
}
\description{
Correlate all columns of two data frames
}
