% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_col.R
\name{corr_col}
\alias{corr_col}
\title{Correlation of columns for tidy data.}
\usage{
corr_col(
  x,
  y = NULL,
  ...,
  xy_join = NULL,
  x_name = "x",
  y_name = "y",
  method = "spearman"
)
}
\arguments{
\item{x}{A tidy data.frame containing numeric data columns.}

\item{y}{(Optional) A tidy data frame containing numeric data columns for pairwise correlation with \code{x}.}

\item{...}{Unused.}

\item{xy_join}{(Optional) A join specification for \code{x} and \code{y} constructed by \link[dplyr]{join_by}.}

\item{x_name}{(Optional) A character of length 1 to be used as the first column name.}

\item{y_name}{(Optional) A character of length 1 to be used as the second column name.}

\item{method}{(Optional) A character of length 1 indicating which correlation coefficient is to be used: \code{"spearman"} (the default), \code{"pearson"}, or \code{"kendall"}.}
}
\value{
A data frame of correlation results.
}
\description{
Correlation of columns for tidy data.
}
