% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_col.R
\name{corr_col}
\alias{corr_col}
\title{Correlations of Columns}
\usage{
corr_col(
  x,
  y,
  options = list(x_join = NULL, y_join = NULL, x_labl = NULL, y_labl = NULL, progress =
    NULL)
)
}
\arguments{
\item{x}{A data frame with a column containing values shared with \code{y}.}

\item{y}{A data frame with a column containing values shared with \code{x}.}

\item{options}{A named list of options to pass to the function.
\itemize{
\item \verb{*_join}: A character scalar of a column name in \code{*} ("x" or "y") to use for joining. Defaults to the first column.
\item \verb{*_labl}: A character scalar describing the inputs ("x" or "y")to label the output columns and metadata. Defaults to "x" or "y".
\item \code{progress}: Boolean value to show function progress. Defaults to \code{FALSE}.
}}
}
\value{
A named list of data frames containing results of correlations between all pairs of numeric columns in \code{x} and \code{y}.
}
\description{
Correlations of Columns
}
