% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_col.R
\name{corr_col}
\alias{corr_col}
\title{Correlations of Columns}
\usage{
corr_col(x, y)
}
\arguments{
\item{x}{A data.frame with a column containing values shared with y in the first position}

\item{y}{A data.frame with a column containing values shared with x in the first position}
}
\value{
A named list of data.frames containing results of correlations between all pairs of numeric columns in x and y
}
\description{
Correlations of Columns
}
