% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_col.R
\name{corr_col_xy}
\alias{corr_col_xy}
\title{(Internal) Two data frame (\code{x}, \code{y}) case of \code{corr_col()}}
\usage{
corr_col_xy(x, y, xy_join = NULL, method)
}
\arguments{
\item{x}{A tidy data frame containing numeric data columns.}

\item{y}{A tidy data frame containing numeric data columns.}

\item{xy_join}{(Optional) A join specification for \code{x} and \code{y} constructed by \link[dplyr]{join_by}.}

\item{method}{(Optional) A character of length 1 indicating which correlation coefficient is to be used: \code{"spearman"} (the default), \code{"pearson"}, or \code{"kendall"}.}
}
\value{
A data frame of correlation results.
}
\description{
(Internal) Two data frame (\code{x}, \code{y}) case of \code{corr_col()}
}
