% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcor_col.R
\name{pcor_col}
\alias{pcor_col}
\title{Partial correlation of columns for tidy data.}
\usage{
pcor_col(
  x,
  y = NULL,
  z = NULL,
  ...,
  xy_join = NULL,
  xz_join = NULL,
  x_name = "x",
  y_name = "y",
  z_name = "z",
  method = "spearman"
)
}
\arguments{
\item{x}{A tidy data.frame containing numeric data columns.}

\item{y}{(Optional) A tidy data frame containing numeric data columns for pairwise correlation with \code{x}.}

\item{z}{(Optional) A tidy data frame containing numeric data columns to use as covariates.}

\item{...}{Unused}

\item{xy_join}{(Optional) A join specification for \code{x} and \code{y} constructed by \link[dplyr]{join_by}.}

\item{xz_join}{(Optional) A join specification for \code{x} and \code{z} constructed by \link[dplyr]{join_by}.}

\item{x_name}{(Optional) A character of length 1 to be used as the first column name.}

\item{y_name}{(Optional) A character of length 1 to be used as the second column name.}

\item{z_name}{(Optional) A character of length 1 to be used as the third column name.}

\item{method}{(Optional) A character of length 1 indicating which correlation coefficient is to be used: \code{"spearman"} (the default), \code{"pearson"}, or \code{"kendall"}.}
}
\value{
A data frame of partial correlation results.
}
\description{
Partial correlation of columns for tidy data.
}
