% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcor_col.R
\name{pcor_col}
\alias{pcor_col}
\title{Partial correlation of two data frames with a third data frame of covariates.}
\usage{
pcor_col(x_data, y_data, z_data, ...)
}
\arguments{
\item{x_data}{A data frame with at least one shared index column and at least one column of numeric data.}

\item{y_data}{A data frame with at least one shared index column and at least one column of numeric data.}

\item{z_data}{A data frame with at least one shared index column and at least one column of numeric data.}

\item{...}{}
}
\value{
A data frame of partial correlation results.
}
\description{
Partial correlation of two data frames with a third data frame of covariates.
}
\examples{
pcor_col(df1, df2, df3)
}
