% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcor_col.R
\name{pcor_col_xyz}
\alias{pcor_col_xyz}
\title{(Internal) Three data frame (\code{x}, \code{y}, \code{z}) case of \code{pcor_col()}}
\usage{
pcor_col_xyz(x, y, z, xy_join, xz_join, method = "spearman")
}
\arguments{
\item{x}{A tidy data frame containing numeric data columns.}

\item{y}{A tidy data frame containing numeric data columns.}

\item{z}{A tidy data frame containing numeric data columns.}

\item{xy_join}{(Optional) A join specification for \code{x} and \code{y} constructed by \link[dplyr]{join_by}.}

\item{xz_join}{(Optional) A join specification for \code{x} and \code{z} constructed by \link[dplyr]{join_by}.}

\item{method}{(Optional) A character of length 1 indicating which correlation coefficient is to be used: \code{"spearman"} (the default), \code{"pearson"}, or \code{"kendall"}.}
}
\value{
A data frame of partial correlation results.
}
\description{
(Internal) Three data frame (\code{x}, \code{y}, \code{z}) case of \code{pcor_col()}
}
